Find an existing Safari WebDriver process or launch a new one
The old code only ever launched a new WebDriver process. The new code
has been refactored to launch a new instance of the web driver if an
existing process cannot be used.

It's also using `async`/`await` and `const` so it will require a fairly
new version of Node now.